+++ mods/sbz_pipeworks/filter_injector.lua
@@ -21,7 +21,7 @@ local function set_filter_formspec(meta)
         fs_helpers.cycling_button(meta, "button[" .. (10.2 - (0.22) - 4) .. ",4.5;4,1", "exmatch_mode",
             { "Exact match - off",
                 "Exact match - on",
-               "Threshold" }) ..
+                "Threshold" }) ..
         pipeworks.fs_helpers.get_inv(6) ..
         "listring[]"
 
@@ -117,36 +117,40 @@ minetest.register_node("pipeworks:automatic_filter_injector", {
         local dir = pipeworks.facedir_to_right_dir(node.param2)
 
         local frompos = vector.subtract(pos, dir)
-        local fromnode = minetest.get_node(frompos)
+        local fromnode = sbz_api.get_node_force(frompos)
 
         if not fromnode then
-            meta:set_string("infotext", "I can't pull from that node - there is no node there?")
+            meta:set_string("infotext", "Can't pull from that node - there is no node there?")
             return 1
         end
 
         local fromdef = minetest.registered_nodes[fromnode.name]
         if not fromdef or not fromdef.tube then
-            meta:set_string("infotext", "I can't pull from that node :/")
+            meta:set_string("infotext", "Can't pull from that node :/")
             return 1
         end
         local fromtube = table.copy(fromdef.tube)
 
         local todir = pipeworks.facedir_to_right_dir(node.param2)
         local topos = vector.add(pos, todir)
-        local tonode = minetest.get_node(topos)
+        local tonode = sbz_api.get_node_force(topos)
+        if not tonode then
+            meta:set_string("infotext", "Can't push to that node - that node does not exist.")
+            return 1
+        end
         local todef = minetest.registered_nodes[tonode.name]
 
         if not todef
             or not (minetest.get_item_group(tonode.name, "tube") == 1
                 or minetest.get_item_group(tonode.name, "tubedevice") == 1
                 or minetest.get_item_group(tonode.name, "tubedevice_receiver") == 1) then
-            meta:set_string("infotext", "I can't push to that node")
+            meta:set_string("infotext", "Can't push to that node")
             return 1
         end
 
         if fromtube then fromtube.input_inventory = fromtube.input_inventory end
         if not (fromtube and fromtube.input_inventory) then
-            meta:set_string("infotext", "I can't pull from that node :/")
+            meta:set_string("infotext", "Can't pull from that node :/")
             return 1
         end
 
@@ -299,7 +303,7 @@ minetest.register_node("pipeworks:automatic_filter_injector", {
                     local start_pos = vector.add(frompos, dir)
 
                     pipeworks.tube_inject_item(pos, start_pos, dir, item,
-                        fakeplayer:get_player_name())
+                        fakeplayer:get_player_name(), nil, topos)
                     return true -- only fire one item, please
                 end
             end